Separate utils module
